# opt: use memcpy and update sizes after memcpy in serialize_batch