optimize `FixedLengthColumnBase<T>::serialize_batch`
- use `memcpy` to replace `strings::memcpy_inlined` when the copied size is `sizeof(T)`.
- update `sizes` after `memcpy`.

----
The information of benchmark test is as follows:
- Dataset: [New York Taxi Data](https://clickhouse.com/docs/en/getting-started/example-datasets/nyc-taxi/).
- The number of rows: 12_9897_9494.
- Query SQL:
``` SQL
-- Q3
SELECT passenger_count, year(pickup_date) AS year_, count(*) 
FROM trips_mergetree 
GROUP BY passenger_count, year_;

-- Q4
SELECT passenger_count, year(pickup_date) AS year_, round(trip_distance) AS distance, count(*) 
FROM trips_mergetree 
GROUP BY passenger_count, year_, distance 
ORDER BY year_, count(*) DESC;
```

`passenger_count` is `int8`, `trip_distance` is `float64`, and the  return type of `year()` is nullable `int32`.



The followings are the cost time before and after the optimization:

Q3:
| Parallel | Before (ms) | After (ms) |
| -------- | ----------- | ---------- |
| 1        | 7323        | 5659       |
| 2        | 3757        | 2960       |
| 4        | 2005        | 1664       |
| 8        | 1463        | 1103       |
| 16       | 1131        | 892        |

Q4:
| Parallel | Before (ms) | After (ms) |
| -------- | ----------- | ---------- |
| 1        | 9751        | 8800       |
| 2        | 4976        | 4517       |
| 4        | 2679        | 2535       |
| 8        | 2048        | 1807       |
| 16       | 1650        | 1468       |

And the cost time rate of `FixedLengthColumnBase<T>::serialize_batch` before and after the optimization are 18.75% and 10.55%, respectively.


</byte-sheet-html-origin>